Fix a deadlock in blob replication
BlobStoreBroadcast.EmitBlob and BlobStoreBroadcast.Register utilize a common
mutex.

WantManager.CreateWants and WantManager.EmitBlob utilize a common mutex.

WantManager.CreateWants calls BlobStoreBroadcast.Register while
BlobStoreBroadcast.EmitBlob calls WantManager.EmitBlob.

This results in a classic deadlock. The simplest way of fixing this was to run
the code which emits notification in a goroutine to make sure that it is
capable of waiting before establishing a lock.